Use unique ID for every allocation
Previously the driver returned memory ID as Buffer Object handle in
function zeMemGetAllocProperties. The Buffer object on close and open
again might assign same handle. This cause the application is unable to
identify the memory. To fix it, this change implements a "static
uint64_t counter" in VPUBufferObject to assign unique ID for every
memory allocation.

Signed-off-by: Wludzik, Jozef <jozef.wludzik@intel.com>